+++ src/main/java/com/example/cowork/repository/TeamRepository.java
@@ -1,8 +1,7 @@
 package com.example.cowork.repository;
 
+import com.example.cowork.entity.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.cowork.entity.Team;
-
-public interface TeamRepository extends JpaRepository<Team, Integer> {
+public interface TeamRepository extends JpaRepository<Team, Long> {
 }
